temp-requirements.txt
flask==3.1.0
flask-cors==5.0.0
tensorflow==2.18.0  # Compatible with Python 3.11
pillow==11.0.0
numpy==2.1.3
requests==2.32.3
gunicorn==23.0.0
